requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
playwright==1.47.0
Pillow==10.4.0
weasyprint==62.3
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.9
supabase>=2.0.0
requests>=2.31.0
resend>=2.0.0
